EnemyTank Invisibility Ultiamte Execution Error
Invisibility ultimate for the enemy not executing anymore after player goes out of range when the enemy was in invisible state

The TargetView was assigned to null, when the player was out of range, and we didnt have a null check on Enter() of EnemyPrimaryAttackState.

That’s why executiong was obstructing.